\name{Semi-package}
\alias{Semi-package}
\alias{Semi}
\docType{Semi}
\title{
\packageTitle{Semi}
}
\description{
\packageDescription{Semi}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Semi}
\packageIndices{Semi}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Semi}

Maintainer: \packageMaintainer{Semi}
}
\references{

}

\keyword{ package }
\seealso{

}
\examples{

}
