% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_sems.R
\name{mel}
\alias{mel}
\title{Takes a value in Hertz and converts it to Mels}
\usage{
mel(x)
}
\arguments{
\item{x}{a numeric value in Hz}
}
\value{
numeric value in Mels
}
\description{
Takes a value in Hertz and converts it to Mels
}
