% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_sems.R
\name{semi}
\alias{semi}
\title{To Semitones}
\usage{
semi(x)
}
\arguments{
\item{x}{a numeric value in Hz}
}
\value{
numeric value in semitones
}
\description{
Takes a value in Hertz and converts it to semitones
}
